share props type
